// frontend/src/client/types.gen.ts
export type UserRead = {
    /**
     * Email
     */
    email: string;
    /**
     * Firstname
     */
    firstName: string;
    /**
     * Lastname
     */
    lastName: string;
    role?: UserRole;
    /**
     * Isactive
     */
    isActive?: boolean;
    /**
     * Id
     */
    id: string;
    /**
     * Created At
     */
    created_at?: string | null;
};

/**
 * UserRole
 */
export type UserRole = 'USER' | 'ADMIN';